- Some clean-up, re-adding some tests that were broken
