+++ src/coin/trx/enum.ts
@@ -9,12 +9,7 @@ export enum ContractType {
   /**
    * This is the multi-sig initialization contract.
    */
-  AccountPermissionUpdate = 'AccountPermissionUpdateContract', 
-}
-
-export enum ContractTypeUrl {
-  TransferContract = 'type.googleapis.com/protocol.TransferContract',
-  AccountPermissionUpdate = 'type.googleapis.com/protocol.AccountPermissionUpdateContract',
+  AccountPermissionUpdate = 'AccountPermissionUpdateContract',
 }
 
 export enum PermissionType {
